Split doctor and local-stack setup out of cli.py
